Movielens 20m result update

+++ experiment/execute.py
@@ -1,3 +1,4 @@
+import os
 import numpy as np
 import pandas as pd
 from tqdm import tqdm
@@ -16,21 +17,33 @@ def execute(train, test, params, model, measure='Cosine', gpu_on=True, analytica
 
     df = pd.DataFrame(columns=columns)
 
-    RQ, Yt, Bias = model(train,
-                         embeded_matrix=np.empty((0)),
-                         iteration=params['iter'],
-                         rank=params['rank'],
-                         lam=params['lambda'],
-                         alpha=params['alpha'],
-                         corruption=params['corruption'],
-                         root=params['root'],
-                         gpu_on=gpu_on)
-    Y = Yt.T
+    if os.path.isfile('{2}/U_{0}_{1}.npy'.format(params['model'], params['rank'], folder)):
 
-    np.save('{2}/U_{0}_{1}'.format(params['model'], params['rank'], folder), RQ)
-    np.save('{2}/V_{0}_{1}'.format(params['model'], params['rank'], folder), Y)
-    if Bias is not None:
-        np.save('{2}/B_{0}_{1}'.format(params['model'], params['rank'], folder), Bias)
+        RQ = np.load('{2}/U_{0}_{1}.npy'.format(params['model'], params['rank'], folder))
+        Y = np.load('{2}/V_{0}_{1}.npy'.format(params['model'], params['rank'], folder))
+
+        if os.path.isfile('{2}/B_{0}_{1}.npy'.format(params['model'], params['rank'], folder)):
+            Bias = np.load('{2}/B_{0}_{1}.npy'.format(params['model'], params['rank'], folder))
+        else:
+            Bias = None
+            
+    else:
+
+        RQ, Yt, Bias = model(train,
+                             embeded_matrix=np.empty((0)),
+                             iteration=params['iter'],
+                             rank=params['rank'],
+                             lam=params['lambda'],
+                             alpha=params['alpha'],
+                             corruption=params['corruption'],
+                             root=params['root'],
+                             gpu_on=gpu_on)
+        Y = Yt.T
+
+        np.save('{2}/U_{0}_{1}'.format(params['model'], params['rank'], folder), RQ)
+        np.save('{2}/V_{0}_{1}'.format(params['model'], params['rank'], folder), Y)
+        if Bias is not None:
+            np.save('{2}/B_{0}_{1}'.format(params['model'], params['rank'], folder), Bias)
 
     progress.subsection("Prediction")
